+++ cloverleaf_ui/public/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>Cloverleaf</title>
+  </head>
+  <body>
+    <script src="./src/index.js" type="text/jsx"></script>
+    <link rel="stylesheet" href="style.css">
+    <div id="root"></div>
+  </body>
+</html>
