+++ cloverleaf_ui/src/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import App from "./components/App";
+
+ReactDOM.render(
+  <App/>,
+  document.getElementById('root')
+);
