+++ html/asset-manifest.json
@@ -1,0 +1,17 @@
+{
+  "files": {
+    "main.js": "/static/js/main.065812fc.chunk.js",
+    "main.js.map": "/static/js/main.065812fc.chunk.js.map",
+    "runtime-main.js": "/static/js/runtime-main.7f2b8839.js",
+    "runtime-main.js.map": "/static/js/runtime-main.7f2b8839.js.map",
+    "static/js/2.cb255655.chunk.js": "/static/js/2.cb255655.chunk.js",
+    "static/js/2.cb255655.chunk.js.map": "/static/js/2.cb255655.chunk.js.map",
+    "index.html": "/index.html",
+    "static/js/2.cb255655.chunk.js.LICENSE.txt": "/static/js/2.cb255655.chunk.js.LICENSE.txt"
+  },
+  "entrypoints": [
+    "static/js/runtime-main.7f2b8839.js",
+    "static/js/2.cb255655.chunk.js",
+    "static/js/main.065812fc.chunk.js"
+  ]
+}
