+++ html/static/js/main.065812fc.chunk.js
@@ -1,0 +1,2 @@
+(this.webpackJsonpcloverleaf_ui=this.webpackJsonpcloverleaf_ui||[]).push([[0],{13:function(e,t,n){"use strict";n.r(t);var s=n(2),r=n(6),a=n.n(r),c=n(1),i=n.n(c),o=n(3),p=n(4),u=n(0);var d=function(){var e,t=s.React.useState(!1),n=Object(p.a)(t,2),r=n[0],a=n[1],c=s.React.useState(!0),d=Object(p.a)(c,2),l=d[0],h=(d[1],s.React.useState(!0)),f=Object(p.a)(h,2),b=f[0];function j(){return(j=Object(o.a)(i.a.mark((function e(){var t,n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return a(!0),e.prev=1,e.next=4,fetch("http://localhost:8888/");case 4:return t=e.sent,e.next=7,t.json();case 7:return n=e.sent,this.session=n.session,console.log("got offer ".concat(n.sdp)),e.next=12,x(n);case 12:e.next=17;break;case 14:e.prev=14,e.t0=e.catch(1),console.error(e.t0);case 17:case"end":return e.stop()}}),e,this,[[1,14]])})))).apply(this,arguments)}function x(e){return y.apply(this,arguments)}function y(){return(y=Object(o.a)(i.a.mark((function t(n){var s,r,a,c=this;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return this.peer=new RTCPeerConnection({}),this.peer.addEventListener("icecandidate",(function(e){return v(c.peer,e)})),this.peer.addEventListener("iceconnectionstatechange",(function(e){return S(c.peer,e)})),this.peer.addEventListener("track",O),t.next=6,e.setRemoteDescription(n);case 6:return t.next=8,e.createAnswer();case 8:return s=t.sent,t.next=11,e.setLocalDescription(s);case 11:return console.log("going to send this answer: ".concat(JSON.stringify(s))),t.next=14,fetch("http://localhost:8888/answer",{method:"POST",body:JSON.stringify({pt:"Answer",payload:JSON.stringify(s),id:"",session:n.session}),headers:{"Content-Type":"application/json"}});case 14:return r=t.sent,t.next=17,r.json();case 17:a=t.sent,console.log("api response is ".concat(a));case 19:case"end":return t.stop()}}),t,this)})))).apply(this,arguments)}function O(e){return g.apply(this,arguments)}function g(){return(g=Object(o.a)(i.a.mark((function e(t){return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:this.clientVideo.current.srcObject=t.streams[0];case 1:case"end":return e.stop()}}),e,this)})))).apply(this,arguments)}function v(e,t){return w.apply(this,arguments)}function w(){return(w=Object(o.a)(i.a.mark((function e(t,n){var s,r,a,c;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(e.prev=0,!n.candidate){e.next=11;break}return e.next=4,fetch("http://localhost:8888/candidate",{method:"POST",body:JSON.stringify({pt:"Candidate",payload:JSON.stringify(n.candidate),id:"",session:this.session}),headers:{"Content-Type":"application/json"}});case 4:return s=e.sent,e.next=7,s.json();case 7:r=e.sent,console.log("api response is ".concat(r)),e.next=19;break;case 11:return console.log("we are done gathering candidates"),e.next=14,fetch("http://localhost:8888/watch",{method:"POST",body:JSON.stringify({pt:"Candidate",payload:JSON.stringify(n.candidate),id:"",session:this.session}),headers:{"Content-Type":"application/json"}});case 14:return a=e.sent,e.next=17,a.json();case 17:c=e.sent,console.log("api response is ".concat(c));case 19:e.next=24;break;case 21:e.prev=21,e.t0=e.catch(0),console.log("error setting ice candidate");case 24:case"end":return e.stop()}}),e,this,[[0,21]])})))).apply(this,arguments)}function S(e,t){return k.apply(this,arguments)}function k(){return(k=Object(o.a)(i.a.mark((function e(t,n){return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:console.log("ice state changed");case 1:case"end":return e.stop()}}),e)})))).apply(this,arguments)}return f[1],Object(s.useRef)(null),Object(u.jsxs)(u.Fragment,{children:[Object(u.jsx)("video",{ref:this.clientVideo,playsinline:!0,autoplay:!0}),Object(u.jsxs)("div",{class:"box",children:[Object(u.jsx)("button",{disabled:r,onClick:function(e){return function(){return j.apply(this,arguments)}()},children:"Start"}),Object(u.jsx)("button",{disabled:l,children:"Call"}),Object(u.jsx)("button",{disabled:b,children:"Hang Up"})]})]})};a.a.render(Object(u.jsx)(d,{}),document.getElementById("root"))}},[[13,1,2]]]);
+//# sourceMappingURL=main.065812fc.chunk.js.map
